+++ src/llm/base.py
@@ -1,0 +1,7 @@
+from loguru import logger
+
+try:
+    from transformers import pipeline
+except Exception as e:
+    pipeline = None
+    logger.warning(f"Transformers indisponível: {e}")
